+++ Search/SearchManager.cs
@@ -16,6 +16,7 @@ public static class SearchManager
 			new PilotMotoSearcher(article, exactArticle, logger),
 			new KtmZipSearcher(article, exactArticle, logger),
 			new MotodartSearcher(article, exactArticle, logger),
+			new MotoFalconSearcher(article, exactArticle, logger),
 		};
 		try
 		{